day07

+++ src/main/java/com/mmns/conmmunity/dto/QuestionDTO.java
@@ -3,18 +3,20 @@ package com.mmns.conmmunity.dto;
 import com.mmns.conmmunity.model.User;
 import lombok.Data;
 
+/**
+ * Created by codedrinker on 2019/5/7.
+ */
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private String tag;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creator;
+    private Long creator;
     private Integer viewCount;
     private Integer commentCount;
     private Integer likeCount;
     private User user;
-
 }
